Add elapsed sensor time column to track `DataFrame`
There does not appear to be a link between the timestamps recorded by the sensor file and the GPS timestamp in the track file, so they will not align when plotting side-by-side.

The sensor data file does contain regular time readings, which can be used to align the sensor readings with UTC time.